# Up to 10x faster scatter.