@vj-krish 2024-02-19
### Performance on M2Max (38-Core GPU) MBP.
#### Before

X (1000000, 64), Indices [(1000000,)]
MLX: 4.529ms
PyTorch: 10.208ms

X (1000000, 64), Indices [(1000000,)]
MLX: 4.630ms
PyTorch: 11.018ms

X (100000, 64), Indices [(100000,)]
MLX: 1.987ms
PyTorch: 1.091ms

X (1000000,), Indices [(1000000,)]
MLX: 0.248ms
PyTorch: 0.290ms

X (20000000,), Indices [(20000000,)]
MLX: 1.110ms
PyTorch: 3.654ms

X (20000000,), Indices [(20000000,)]
MLX: 5.488ms
PyTorch: 7.152ms

X (1000000, 64), Indices [(1000000,)]
MLX: 4.537ms
PyTorch: 10.634ms

X (10000000, 64), Indices [(10000000,)]
MLX: 43.065ms
PyTorch: 100.996ms

X (1000, 10000, 64), Indices [(1000,)]
MLX: 58.935ms
PyTorch: 108.575ms

X (10000, 100, 100, 21), Indices [(10000,)]
MLX: 235.824ms
PyTorch: 336.351ms

X (1000, 10), Indices [(1000,), (1000,)]
MLX: 0.438ms
PyTorch: 0.251ms

#### After

X (1000000, 64), Indices [(1000000,)]
MLX: 0.837ms
PyTorch: 10.298ms

X (1000000, 64), Indices [(1000000,)]
MLX: 1.242ms
PyTorch: 11.068ms

X (100000, 64), Indices [(100000,)]
MLX: 1.740ms
PyTorch: 1.074ms

X (1000000,), Indices [(1000000,)]
MLX: 0.231ms
PyTorch: 0.284ms

X (20000000,), Indices [(20000000,)]
MLX: 0.964ms
PyTorch: 3.655ms

X (20000000,), Indices [(20000000,)]
MLX: 5.471ms
PyTorch: 7.195ms

X (1000000, 64), Indices [(1000000,)]
MLX: 0.829ms
PyTorch: 10.707ms

X (10000000, 64), Indices [(10000000,)]
MLX: 6.666ms
PyTorch: 101.062ms

X (1000, 10000, 64), Indices [(1000,)]
MLX: 10.013ms
PyTorch: 108.335ms

X (10000, 100, 100, 21), Indices [(10000,)]
MLX: 22.924ms
PyTorch: 335.587ms

X (1000, 10), Indices [(1000,), (1000,)]
MLX: 0.422ms
PyTorch: 0.263ms