@TristanBilot 2024-02-19
Wow, impressive work again @vj-krish 💪
I'll give it a try on real datasets soon!