@vj-krish 2024-02-20
> Looks awesome!! @vj-krish can you check the failed tests?

Should pass now. Added some host side code to bail from the fast path in a couple of scenarios where linear addressing isn't applicable.