@awni 2024-02-20
I think this is broken for col contiguous inputs but maybe I am wrong about that. Might be good to add a test where you send in transposed inputs.