@vj-krish 2024-02-20
> > I think maybe what you want is all the arrays are row_contiguous (e.g. inputs[i].flags().row_contiguous == True).
> 
> Did you change this?

Not yet. Was waiting for the thread to converge before addressing it.