@awni 2024-02-20
I see. Ok well the indices are 1D so the check you have now is fine but it's simpler/faster to just check `row_contiguous` 

For the update if it is multidimensional and col contiguous I think (?) it is a bug